vpr: Increased the maximum size of the feasible block array
The extra carry chain architecture have more primitives in the logic
block than other architectures. Therefore the maximum size of the
feasible block array needs to be increased to be able to absorb all
two pin nets in the cluster (e.g. logical not of subtractors or
output registers connected to the adder outputs)